+++ grid2op/tests/test_n_busbar_per_sub.py
@@ -536,13 +536,9 @@ class TestAction_3busbars(unittest.TestCase):
         assert "change_bus" not in type(self.env.action_space()).authorized_keys
         assert not self.env.action_space.supports_type("change_bus")
     
-    def test_shunt(self):
-        el_id = 0
-        bus_val = -1
+    def _aux_test_action_shunt(self, act, el_id, bus_val):
         name_xxx = None
         el_nms = None
-        
-        act = self.env.action_space({"shunt": {"set_bus": [(0, -1)]}})
         # self._aux_test_action(act, type(self.env).name_shunt, el_id, bus_val, None)  # does not work for a lot of reasons
         assert not act.is_ambiguous()[0]
         tmp = f"{act}"  # test the print does not crash
@@ -550,8 +546,21 @@ class TestAction_3busbars(unittest.TestCase):
         self._aux_test_act_consistent_as_dict(tmp, name_xxx, el_id, bus_val)
         tmp = act.as_serializable_dict()  # test I can convert to another type of dict
         self._aux_test_act_consistent_as_serializable_dict(tmp, el_nms, el_id, bus_val)
-    
-    
+        
+    def test_shunt(self):
+        el_id = 0
+        bus_val = -1
+        act = self.env.action_space({"shunt": {"set_bus": [(0, bus_val)]}})
+        self._aux_test_action_shunt(act, el_id, bus_val)
+        
+        for bus_val in range(type(self.env).n_busbar_per_sub):
+            act = self.env.action_space({"shunt": {"set_bus": [(0, bus_val)]}})
+            self._aux_test_action_shunt(act, el_id, bus_val)
+            
+        act = self.env.action_space()
+        with self.assertRaises(IllegalAction):
+            act = self.env.action_space({"shunt": {"set_bus": [(0, type(self.env).n_busbar_per_sub + 1)]}})
+
 class TestAction_1busbar(TestAction_3busbars):
     def get_nb_bus(self):
         return 1